Add UUID for client IP address for FormsSignIn QFlow API session authentication requires unique IP address value, otherwise existing sessions for the same username are invalidated when attempting to sign in again.
From Simon Pearson (NEXA):

TL;DR Explanation

It is because you are not changing the IPAddress field value when you call FormsSignIn - you are leaving this as NULL each time.

Q–Flow will assume that you are the same person, and allow you straight in (including issuing you a new APISessionId).

When calling a Q-Flow web service, you pass through the APISessionId that was generated by the FormsSignIn service (which is also stored against the User record in a field).  The method you are calling with the APISessionId will try to authenticate you.