Return correct number of available times for today
Previously, AvailableDatesController didn't return a correct number of
available appointment times for today, since the count included unfilled
appointment times from before current time. (E.g., a query that took
place at 1pm would include in its count of slots the unused 10am slot,
even though this clearly can no longer be booked.)

Now, AvailableTimesService is used to fetch the correct number of
available appointment times for today, if today comes back with
availability according to AvailableDatesService.

It also returns the actual times for today to the client, so the client
doesn't need to fetch the times for today again from the server.
Currently our client doesn't make use of this.

By the way, 'today' is the date according to the current time of the
appointment venue (not according to server time).